Правка 1 задание 5.12 ветка module5-task1 от 09.06.2021
