+++ index.html
@@ -230,6 +230,6 @@
     </div>
   </template>
   <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-  <script src="js/main.js" defer type="module"></script>
+  <script src="./js/main.js" defer type="module"></script>
 </body>
 </html>
